+++ src/bika/lims/browser/analysisrequest/add2.py
@@ -1699,12 +1699,21 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
 
         return self.handle_redirect(ARs.values(), message)
 
+    def is_automatic_label_printing_enabled(self):
+        """Returns whether the automatic printing of barcode labels is active
+        """
+        setup = api.get_setup()
+        auto_print = setup.getAutoPrintStickers()
+        auto_receive = setup.getAutoreceiveSamples()
+        action = "receive" if auto_receive else "register"
+        return action in auto_print
+
     def handle_redirect(self, uids, message):
         """Handle redirect after sample creation or cancel
         """
         # Automatic label printing
         setup = api.get_setup()
-        auto_print = setup.getAutoPrintStickers()
+        auto_print = self.is_automatic_label_printing_enabled()
         immediate_results_entry = setup.getImmediateResultsEntry()
         redirect_to = self.context.absolute_url()
 
@@ -1724,11 +1733,9 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
                         ",".join(uids),  # copy_from
                         len(uids),  # ar_count
                         sample_uids)  # sample_uids
-        elif "register" in auto_print and sample_uids:
-            redirect_to = "{}/sticker?autoprint=1&template={}&items={}".format(
-                self.context.absolute_url(),
-                setup.getAutoStickerTemplate(),
-                sample_uids)
+        elif auto_print and sample_uids:
+            redirect_to = "{}/sticker?autoprint=1&items={}".format(
+                self.context.absolute_url(), sample_uids)
         elif immediate_results_entry and sample_uids:
             redirect_to = "{}/multi_results?uids={}".format(
                 self.context.absolute_url(),