+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 ------------------
 
 - #2120 Support for Multiple component analysis
+- #2190 Fix sample actions without translation
+- #2189 Fix auto-print of barcode labels when auto-receive is enabled
+- #2188 Fix archive path for Generic Setup file imports
+- #2185 Fix analysis profiles not recognised on AR template creation
+- #2186 Fix sticker template for sample type is not selected by default
+- #2183 Fix instrument supplier does not load on test data import
 - #2180 Rely on analysis permission when displaying results
 - #2179 Fix Traceback when removing a Worksheet
 - #2178 AT Queryselect Widget